fixed custom exception

+++ src/main/java/com/backend/ecommercebackend/authentication/service/impl/AuthenticationServiceImpl.java
@@ -18,7 +18,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -27,6 +26,7 @@ import org.springframework.stereotype.Service;
 
 import java.io.IOException;
 import java.time.LocalDateTime;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -49,18 +49,21 @@ public class AuthenticationServiceImpl implements AuthenticationService {
         }
 
         if(request.getAcceptTerms().equals(Boolean.FALSE)){
-            throw new RuntimeException("accept terms can be choose");
+            throw new ApplicationException(Exceptions.TERMS_ACCEPTANCE_EXCEPTION);
         }
+
+        Optional<User> existingUser = repository.findByEmail(request.getEmail());
+        if (existingUser.isPresent()) {
+            throw new ApplicationException(Exceptions.USER_ALREADY_EXIST);
+        }
+
         User user = authMapper.RegisterDtoToEntity(request,passwordEncoder);
         user.setRole(Role.USER);
         user.setCreatedAt(LocalDateTime.now());
         user.setUpdatedAt(LocalDateTime.now());
-        try {
-            repository.save(user);
-        }
-        catch (DataIntegrityViolationException e) {
-            throw new ApplicationException(Exceptions.USER_ALREADY_EXIST);
-        }
+
+        repository.save(user);
+
         emailService.deleteStoredEmail(request.getEmail());
 
         String accessToken = jwtService.generateAccessToken(user.getEmail());
